# api/src/main/resources/openapi.yaml
openapi: 3.0.4
info:
  title: Swagger Transactions - OpenAPI 3.0
  description: Swagger for Transactions App
  version: 1.0.0
tags:
  - name: transactions
    description: Everything about CRUD operations related to transactions
paths:
  /transactions:
    post:
      tags:
        - transactions
      summary: Create a new transaction.
      description: Create a new transaction.
      operationId: createTransaction
      requestBody:
        $ref: '#/components/requestBodies/Transaction'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Invalid input
        default:
          description: Unexpected error
      security:
        - openIdConnect:
            - SCOPE_reader
            - SCOPE_writer

    get:
      tags:
        - transactions
      summary: Finds all transactions.
      description: Finds all transactions.
      operationId: getAll
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
        default:
          description: Unexpected error
      security:
        - openIdConnect:
            - SCOPE_reader
            - SCOPE_writer
  /transactions/{id}:
    get:
      tags:
        - transactions
      summary: Find a transaction by ID.
      description: Returns a single transaction.
      operationId: getTransaction
      parameters:
        - name: id
          in: path
          description: ID of transaction to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Invalid ID supplied
        '404':
          description: Transactions not found
        default:
          description: Unexpected error
      security:
        - openIdConnect:
            - SCOPE_reader
            - SCOPE_writer

    put:
      tags:
        - transactions
      summary: Updates a transaction.
      description: Updates a transaction.
      operationId: updateTransaction
      requestBody:
        $ref: '#/components/requestBodies/Transaction'
      parameters:
        - name: id
          in: path
          description: ID of transaction that needs to be updated
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Invalid input
        default:
          description: Unexpected error
      security:
        - openIdConnect:
            - SCOPE_reader
            - SCOPE_writer
    patch:
      tags:
        - transactions
      summary: Patch a transaction.
      description: Patch a transaction.
      operationId: patchTransaction
      parameters:
        - name: id
          in: path
          description: ID of transaction that needs to be patched
          required: true
          schema:
            type: string
      requestBody:
        $ref: '#/components/requestBodies/JsonPatch'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Invalid input
        default:
          description: Unexpected error
      security:
        - openIdConnect:
            - SCOPE_reader
            - SCOPE_writer
    delete:
      tags:
        - transactions
      summary: Deletes a transaction.
      description: Delete a transaction.
      operationId: deleteTransaction
      parameters:
        - name: id
          in: path
          description: Pet id to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction deleted
        '400':
          description: Invalid transaction value
        default:
          description: Unexpected error
      security:
        - openIdConnect:
            - SCOPE_reader
            - SCOPE_writer

components:
  schemas:
    BankAccount:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          example: 10
        category:
          type: string
        label:
          type: string
    Category:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          example: 10
        category:
          type: string
        label:
          type: string
        type:
          type: string
          enum:
            - FIXE
            - COURANTE
            - EXTRA
    TransactionDetail:
      type: object
      properties:
        category:
          $ref: '#/components/schemas/Category'
        income:
          type: number
          format: float
        outcome:
          type: number
          format: float
        description:
          type: string
        bankAccount:
          $ref: '#/components/schemas/BankAccount'
        cost:
          type: integer
          format: int64
        costAbs:
          type: integer
          format: int64
    Transaction:
      required:
        - date
        - transactions
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDetail'
        date:
          type: string
          format: date-time
        cost:
          type: integer
          format: int64
        costAbs:
          type: integer
          format: int64
        type:
          type: string
          enum:
            - INCOME
            - OUTCOME
            - TRANSFER

    JsonPatch:
      type: array
      items:
        $ref: "#/components/schemas/PatchDocument"

    PatchDocument:
      type: object
      description: A JSONPatch document as defined by RFC 6902
      required:
        - "op"
        - "path"
      properties:
        op:
          type: string
          description: The operation to be performed
          enum:
            - "add"
            - "remove"
            - "replace"
            - "move"
            - "copy"
            - "test"
        path:
          type: string
          description: A JSON-Pointer
        value:
          type: object
          description: The value to be used within the operations.
        from:
          type: string
          description: A string containing a JSON Pointer value.
  requestBodies:
    Transaction:
      description: Transaction object that needs to be added.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Transaction'
    JsonPatch:
      content:
        application/json-patch+json:
          schema:
            $ref: '#/components/schemas/JsonPatch'

  securitySchemes:
    openIdConnect:
      type: openIdConnect
      openIdConnectUrl: https://keycloak/.well-known/openid-configuration